No public description
PiperOrigin-RevId: 861002357
