--- src/parallax/sharding/ddp_test.py
@@ -1,55 +1,0 @@
-"""Tests for DDP sharding."""
-
-import os
-
-from absl.testing import absltest
-from absl.testing import parameterized
-from flax import nnx
-import jax
-import jax.numpy as jnp
-from parallax.examples import models
-from parallax.sharding import ddp
-
-NamedSharding = jax.sharding.NamedSharding
-P = jax.sharding.PartitionSpec
-
-
-class DdpTest(parameterized.TestCase):
-
-  def setUp(self):
-    super().setUp()
-    self.mesh = jax.make_mesh(
-        (4, 2),
-        ('data', 'model'),
-        axis_types=(jax.sharding.AxisType.Auto,) * len(('data', 'model')),
-    )
-    self.enter_context(jax.set_mesh(self.mesh))
-
-  def test_get_shardings(self):
-    model = models.SimpleMLP(16, 64, 16, rngs=nnx.Rngs(0))
-    inputs = (jnp.ones((4, 16)),)
-    graphdef, state = nnx.split(model)
-
-    def fn(state, *inputs):
-      return nnx.merge(graphdef, state)(*inputs)
-
-    (params_assignments, inputs_assignments), output_assignments = (
-        ddp.get_shardings(fn, state, *inputs)
-    )
-
-    expected_params = {
-        'layers': {
-            0: {'bias': P(None), 'kernel': P(None, None)},
-            1: {'bias': P(None), 'kernel': P(None, None)},
-        }
-    }
-
-    self.assertEqual(nnx.to_pure_dict(params_assignments), expected_params)
-    self.assertEqual(inputs_assignments, P('data', None))
-    self.assertEqual(output_assignments, P('data', None))
-
-
-if __name__ == '__main__':
-  # Fake 8 CPUs for testing.
-  os.environ['XLA_FLAGS'] = '--xla_force_host_platform_device_count=8'
-  absltest.main()
